fix a bug in spectral feature computation
I wanted to validate that the application of raman and IR intensities is done correctly and realized that always k_width values are the same if the bin_width is small than k_width. This was caused by a missing float() casting. Should not have any influence on the results because we always used bin_width = 1.5 sigma